add percentage input element
