+++ src/hello-world/src/cci_components/cci_part/CCiLabSetupOperation.js
@@ -8,137 +8,7 @@ import { dividerCCS, isValidString, isValidValue, getRandomInt } from "./CCiLabU
 import {DateInput} from "./CCiLabDateInput"
 import {NumberInput} from "./CCiLabNumberInput"
 import {TextInput} from "./CCiLabTextInput"
-
-const SetupComponentOp=(props)=>{
-  const { t } = useTranslation(['operations','commands'], {useSuspense: false});
-
-  let inputValue = (props.value === null)? '': props.value;
-
-  let inputClassName = 'text-primary m-0 p-0 border-0 cursor-pointer';
-  let cellWidth = (typeof props.cellCnt !== 'undefined' && props.cellCnt === 1) ?  '20rem' : '10rem';
-  let inputStyle={'backgroundColor': `${styles.cciBgColor}`, width: `${cellWidth}`};
-  let inputType='text';
-  let tooltipOnMode=['click','hover'];
-  let tooltipPosition='top left';
-  let inputName=props.title;
-  let appendPercentage = props.title.includes('-rate')  ? true : false;
-
-  let rateInputElement = React.createRef();
-
-  if( appendPercentage )
-  {
-    let value =  parseFloat(inputValue);
-    if( isNaN( value ) )
-      inputValue='';
-    else
-      inputValue = value + '(%)';
-  }
-
-
-  const [input, setInput] = useState(`${inputValue}`); // '' is the initial state value
-
-  const filterInputValue=( e )=>{
-      // https://stackoverflow.com/questions/10023845/regex-in-javascript-for-validating-decimal-numbers
-      // https://regexr.com/ test expression
-      setInput(e.target.value);
-  };
-
-  const rateAppendPercentage=(_value)=>{
-    let value=parseFloat(_value);
-    if( isNaN(value) )
-      value='';
-    else
-      value += value ? ' (%)' : '';
-
-    setInput(value);
-  };
-
-  const onBlurHandler=(e)=>{
-    rateAppendPercentage(e.target.value);
-  };
-
-  // https://blog.bitsrc.io/understanding-currying-in-javascript-ceb2188c339
-  const updateValue=(props)=>(e)=>{
-    //   if( typeof props.handler !== 'undefined')
-    //   {
-    //     if( typeof e.target !== 'undefined' && e.target.value === '' && props.title ==='part-name')
-    //       e.target.value = 'add-part';
-
-    //     console.log("SetupComponentOp - updateValue: " + e.target.value);
-
-    //     props.handler(e.target.value, props.component);
-    //   }
-    onUpdateValueEnterKey( props, e.target);
-  }
-
-  const onUpdateValueEnterKey=(props, target )=>{
-    if( typeof props.handler !== 'undefined')
-    {
-      if( typeof target !== 'undefined' && target.value === '' && props.title ==='part-name')
-        target.value = 'add-part';
-
-      console.log("SetupComponentOp - updateValue: " + target.value);
-
-      props.handler(props.id, target.value, props.component);
-
-    }
-  }
-
-  const enterKeyHandler=(e)=>{
-    if( typeof e.key !== 'undefined' && e.key ==='Enter')
-    {
-      if( appendPercentage )
-      {
-        rateAppendPercentage(rateInputElement.current.value);
-      }
-      else
-      {
-        onUpdateValueEnterKey(props, e.target);
-      }
-    }
-  };
-
-
-
-  // https://medium.freecodecamp.org/reactjs-pass-parameters-to-event-handlers-ca1f5c422b9
-  return (
-    <div className='d-flex justify-content-between'
-         style={{backgroundColor: `${styles.cciBgColor}`}}>
-          <Popup
-              trigger={
-                <input className={`${inputClassName}`}
-                      ref={ appendPercentage? rateInputElement : null }
-                      id={inputName}
-                      type={`${inputType}`}
-                      style={inputStyle}
-                      placeholder={t(`operations:${props.title}`)}
-                      name={inputName}
-                      value={ input }
-                      min = { inputType.includes('number') ? 1 : null}
-                      onChange={ updateValue(props) }
-                      onClose={updateValue(props)}
-                      onInput={(e)=>{filterInputValue(e)}}
-                      onKeyPress={ (e)=>enterKeyHandler(e) }
-                      onBlur={ (e)=>{onBlurHandler(e)} }/>
-              }
-              id={`${props.component.displayLogic.key}-tooltip`}
-              position={tooltipPosition}
-              closeOnDocumentClick
-              on={tooltipOnMode}
-              arrow={true}
-              arrowStyle={{backgroundColor: 'white'}}
-              mouseLeaveDelay={0}
-              mouseEnterDelay={0}
-              contentStyle={{ padding: '0px'}}
-              >
-              <div className='text-nowrap m-0 px-1'>
-                {t(`operations:${props.title}`)}
-              </div>
-          </Popup>
-    </div>
-  );
-}
-
+import {PercentageInput} from "./CCiLabPercentageInput"
 
 export const initializeOp=( component )=>{
   let operation= JSON.parse(sessionStorage.getItem(`${component.displayLogic.key}_${component.businessLogic.name}_op`)) || _initializeOp();
@@ -739,10 +609,11 @@ export const SetupOP=(props)=>{
                     component={props.component}
                     handler={setSetupCost}/>
                 <hr className={dividerCCS.hDividerClassName }  style={dividerCCS.vDividerStyle}/>    
-                <SetupComponentOp
+                <PercentageInput
                   title='scrap-rate'
                   id={-1}
                   cellCnt={2}
+                  mrpInputType='operations'
                   value={props.component.operation.scrapRate }
                   component={props.component}
                   handler={setScrapRate}/>
